backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
httpx==0.25.1
pillow==10.0.1
boto3==1.29.7
openai==1.3.5
pandas==2.1.3
numpy==1.26.4
python-crontab==3.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
tiktoken==0.7.0
playwright==1.45.0
